Move fn unevaluated from syntax.rs to helpers.rs
Another small reorganization of the helper methods.
