+++ src/functions/list_helpers_ast/mod.rs
@@ -6,10 +6,9 @@
 use crate::InterpreterError;
 use crate::helpers::{
   bool_expr, call, call0, call1, div2, minus2, neg1, plus2, pow2, times2,
+  unevaluated,
 };
-use crate::syntax::{
-  BinaryOperator, ComparisonOp, Expr, UnaryOperator, unevaluated,
-};
+use crate::syntax::{BinaryOperator, ComparisonOp, Expr, UnaryOperator};
 use num_bigint::{BigInt, Sign};
 
 mod aggregation;
